moved the retry mechanism out of LibGit2Sharp classlib  f2b613
