+++ src/GitVersion.Core/Model/Exceptions/LockedFileException.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace GitVersion
+{
+    public class LockedFileException : Exception
+    {
+        public LockedFileException(Exception inner) : base(inner.Message, inner)
+        {
+        }
+    }
+}
